Reorganizing api classes and tests
