+++ cyoa/app/models/api/nws_request.rb
