+++ cyoa/app/models/api/parameters/cloud_cover_parameter.rb
