+++ cyoa/app/models/api/parameters/conditions_icon_parameter.rb
