+++ cyoa/app/models/api/parameters/location_parameter.rb
