+++ cyoa/app/models/api/parameters/temperature_type_parameter.rb
