+++ cyoa/app/models/api/parameters/weather_conditions_parameter.rb
